+++ packages/host/src/index.ts
@@ -3,5 +3,7 @@ export {
     isInsideContainerSync,
     getHostLocalStorage,
     getHostProvider,
+    getStatementStore,
 } from "./container.js";
-export type { HostLocalStorage } from "./types.js";
+export type { HostLocalStorage, HostStatementStore } from "./types.js";
+export { BULLETIN_RPCS, DEFAULT_BULLETIN_ENDPOINT } from "./chains.js";