+++ examples/t3rminal-lite/CHANGELOG.md
@@ -1,5 +1,23 @@
 # t3rminal-lite
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [652b215]
+- Updated dependencies [652b215]
+  - @polkadot-apps/signer@1.1.0
+  - @polkadot-apps/chain-client@2.0.2
+  - @polkadot-apps/tx@0.3.5
+
+## 0.0.9
+
+### Patch Changes
+
+- @polkadot-apps/chain-client@2.0.1
+- @polkadot-apps/signer@1.0.1
+- @polkadot-apps/tx@0.3.4
+
 ## 0.0.8
 
 ### Patch Changes