+++ packages/chain-client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @polkadot-apps/chain-client
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [652b215]
+  - @polkadot-apps/host@0.5.1
+
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [c83fcfa]
+  - @polkadot-apps/host@0.5.0
+
 ## 2.0.0
 
 ### Major Changes