+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @polkadot-apps/storage
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies [652b215]
+  - @polkadot-apps/host@0.5.1
+
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [c83fcfa]
+  - @polkadot-apps/host@0.5.0
+
 ## 0.2.6
 
 ### Patch Changes